requirements.txt
# 環境変数の設定用
python-dotenv

# LangChain用
langchain
langchain-community
langchain-openai

# retrieval用
pypdf 
chromadb